fix build tags for traceroute runner
